// lib/api/generated/models/exerciseRole.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CKS Slavia API
 * API panelu klubowego CKS Slavia Ruda Śląska
 * OpenAPI spec version: 2.2.0.1+1
 */

export type ExerciseRole = typeof ExerciseRole[keyof typeof ExerciseRole];


export const ExerciseRole = {
  warmup: 'warmup',
  main: 'main',
  accessory: 'accessory',
} as const;