// lib/api/generated/models/planOrigin.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CKS Slavia API
 * API panelu klubowego CKS Slavia Ruda Śląska
 * OpenAPI spec version: 2.2.0.1+1
 */

export type PlanOrigin = typeof PlanOrigin[keyof typeof PlanOrigin];


export const PlanOrigin = {
  wizard: 'wizard',
  catalog: 'catalog',
  ai: 'ai',
  manual: 'manual',
} as const;