// lib/api/generated/models/sessionTemplate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CKS Slavia API
 * API panelu klubowego CKS Slavia Ruda Śląska
 * OpenAPI spec version: 2.2.0.1+1
 */

export type SessionTemplate = typeof SessionTemplate[keyof typeof SessionTemplate];


export const SessionTemplate = {
  snatch: 'snatch',
  clean_jerk: 'clean_jerk',
  squat: 'squat',
  mixed: 'mixed',
  light: 'light',
  custom: 'custom',
} as const;